pyproject.toml
[project]
name = "consiglio"
version = "0.1.0"
description = "CLI bargaining-forecast engine inspired by strategic actor modeling"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
consiglio = "consiglio.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["consiglio"]
